Updating notice for Sublime Text usage policy
Although Sublime Text has a 'Pay for continued use' price policy, the 'not free' notice is quite misleading since users are able to download, and use/test the Editor for free without any limitation apart from the regular 'Please consider buying a license ...'.

Once they become familiar with the tool and start using it for commercial purposes, they can buy a license, which in my opinion is really worth the money. It is a great tool, and I've been using it for many years now as my main code editor, and I think it's a little bit unfair to put the 'not free' notice, which might discourage our students from giving it a try.