Update MAKEME.md
This might be up for debate, but I suggest we take out the assignment to provide feedback on each other's code. I don't think it adds much to the skillset of beginning students to provide feedback as they hardly know how to code themselves. Especially when teachers are giving feedback already, it seems a bit unnecessary. 

I also changed the deadline for the homework assignment from Saturday to Friday.

Finally, I suggest moving the literature from Step 1 to the Readme.md -- why are reading materials mentioned in the homework file when we have a readme file? :)